refactor(direct-control): extract capability catalog schemas
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only Civ7CapabilityCatalogEntrySchema, Civ7CapabilityCatalogSchema, and their derived catalog entry/result types into src/catalog/capabilities.ts while keeping public facade re-exports in index.ts.

Preserves TypeBox schema shape and public catalog result typing. Broader public constants/types, procedure schemas, operation/proof telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- public-api.test.ts runtime-and-catalog.test.ts
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- git diff --check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run resources:status
- bun run check:cli
- bun run test:cli:play

Local package/schema relocation proof only; no runtime, live-game, AI/hotseat product-path, semantic CLI, telemetry, or procedure-core proof claimed.
