+++ packages/civ7-direct-control/src/index.ts
@@ -2,8 +2,6 @@ import { readdir, readFile, stat } from "node:fs/promises";
 import { homedir } from "node:os";
 import { basename, extname, join, resolve } from "node:path";
 import { Socket, createConnection } from "node:net";
-import { Type, type Static } from "typebox";
-import { Value } from "typebox/value";
 import {
   assertCiv7ComponentId,
   Civ7ComponentIdSchema,
@@ -33,9 +31,13 @@ import {
   type FreshLogMarkerProof,
 } from "./proof/log-markers.js";
 import {
+  Civ7CapabilityCatalogEntrySchema,
+  Civ7CapabilityCatalogSchema,
   createStaticCiv7CapabilityCatalog as createStaticCiv7CapabilityCatalogFromModule,
   generateCiv7CapabilityCatalog as generateCiv7CapabilityCatalogFromModule,
   loadCiv7OfficialResourceCapabilities,
+  type Civ7CapabilityCatalog,
+  type Civ7CapabilityCatalogEntry,
 } from "./catalog/capabilities.js";
 import {
   appUiSnapshotFromCommandResult,
@@ -118,6 +120,14 @@ export type {
   FreshLogMarkerProof,
 } from "./proof/log-markers.js";
 export { loadCiv7OfficialResourceCapabilities } from "./catalog/capabilities.js";
+export {
+  Civ7CapabilityCatalogEntrySchema,
+  Civ7CapabilityCatalogSchema,
+} from "./catalog/capabilities.js";
+export type {
+  Civ7CapabilityCatalog,
+  Civ7CapabilityCatalogEntry,
+} from "./catalog/capabilities.js";
 
 export const DEFAULT_CIV7_TUNER_HOST = "127.0.0.1";
 export const DEFAULT_CIV7_TUNER_PORT = 4318;
@@ -2053,40 +2063,6 @@ export type Civ7DestinationAnalysisResult = Readonly<{
   pointsOfInterest: unknown;
   notes: ReadonlyArray<string>;
 }>;
-
-export const Civ7CapabilityCatalogEntrySchema = Type.Object({
-  id: Type.String(),
-  name: Type.String(),
-  role: Type.Union([Type.Literal("app-ui"), Type.Literal("tuner"), Type.Literal("shared")]),
-  kind: Type.Union([
-    Type.Literal("root"),
-    Type.Literal("method"),
-    Type.Literal("read-wrapper"),
-    Type.Literal("action-wrapper"),
-    Type.Literal("enum"),
-    Type.Literal("gameinfo-table"),
-  ]),
-  owner: Type.String(),
-  risk: Type.Union([Type.Literal("read"), Type.Literal("low"), Type.Literal("medium"), Type.Literal("high")]),
-  provenance: Type.Array(Type.String()),
-  state: Type.Optional(Type.String()),
-  root: Type.Optional(Type.String()),
-  method: Type.Optional(Type.String()),
-  wrapper: Type.Optional(Type.String()),
-  confidence: Type.Union([Type.Literal("source"), Type.Literal("recorded-live-proof"), Type.Literal("runtime"), Type.Literal("inference")]),
-  description: Type.Optional(Type.String()),
-});
-
-export type Civ7CapabilityCatalogEntry = Static<typeof Civ7CapabilityCatalogEntrySchema>;
-
-export const Civ7CapabilityCatalogSchema = Type.Object({
-  generatedAt: Type.String(),
-  source: Type.Union([Type.Literal("runtime"), Type.Literal("static"), Type.Literal("merged")]),
-  version: Type.String(),
-  entries: Type.Array(Civ7CapabilityCatalogEntrySchema),
-});
-
-export type Civ7CapabilityCatalog = Static<typeof Civ7CapabilityCatalogSchema>;
 
 export type Civ7CapabilityCatalogOptions = Civ7DirectControlOptions & Readonly<{
   includeRuntime?: boolean;
